Split MainFrame::Update() functions into its own file
